Updating infrastructure to better support moonorepo
This commit refactors the GitHub Actions workflow to make it more DRY (Don't
Repeat Yourself) and adhere to best practices. The main changes include:
- Refactored .editorconfig and .ignore
- Move environment variables to the top-level `env` section
  - Allows for easier updating and reuse of versions throughout the workflow
  - Removes the need for redundant `env` sections in individual jobs

- Simplify step names and remove redundant comments
  - Improves readability and maintainability of the workflow
  - Focuses on the essential information for each step

- Ensure consistent indentation and formatting
  - Enhances the overall readability and organization of the workflow
  - Follows best practices for YAML syntax and structure

The refactored workflow now triggers on pushes and pull requests to all branches,
providing comprehensive CI coverage for the project. It sets up PHP, Python, and
Terraform with the specified versions, and runs various checks and tests, including:

- PHP linting
- PHP code style checks using phpcs with PSR-12 standard
- PHP unit tests using phpunit
- Ansible linting using ansible-lint

Finally, the workflow initializes, validates, and applies the Terraform configuration
to provision the necessary infrastructure.

This refactoring effort aims to improve the maintainability, readability, and
effectiveness of the GitHub Actions workflow for the WordPress project.